requirements.txt
PyMuPDF
Pillow
opencv-python
requests
tkhtmlview
markdown
